+++ TestTools/StreamChatTestTools/Mocks/StreamChat/Repositories/ChannelRepository_Mock.swift
@@ -9,7 +9,11 @@
 @testable import StreamChat
 import Foundation
 
-class ChannelRepository_Mock: ChannelRepository {
+class ChannelRepository_Mock: ChannelRepository, Spy {
+    let spyState = SpyState()
+    var getChannel_store: Bool?
+    var getChannel_result: Result<ChatChannel, Error>?
+    
     var markReadCid: ChannelId?
     var markReadUserId: UserId?
     var markReadResult: Result<Void, Error>?
@@ -20,7 +24,14 @@ class ChannelRepository_Mock: ChannelRepository {
     var markUnreadLastReadMessageId: UserId?
     var markUnreadResult: Result<ChatChannel, Error>?
 
+    override func getChannel(for query: ChannelQuery, store: Bool, completion: @escaping (Result<ChatChannel, any Error>) -> Void) {
+        record()
+        getChannel_store = store
+        getChannel_result?.invoke(with: completion)
+    }
+    
     override func markRead(cid: ChannelId, userId: UserId, completion: ((Error?) -> Void)? = nil) {
+        record()
         markReadCid = cid
         markReadUserId = userId
 
@@ -30,6 +41,7 @@ class ChannelRepository_Mock: ChannelRepository {
     }
 
     override func markUnread(for cid: ChannelId, userId: UserId, from messageId: MessageId, lastReadMessageId: MessageId?, completion: ((Result<ChatChannel, Error>) -> Void)? = nil) {
+        record()
         markUnreadCid = cid
         markUnreadUserId = userId
         markUnreadMessageId = messageId