+++ Tests/StreamChatTests/Workers/Background/ConnectionRecoveryHandler_Tests.swift
@@ -500,7 +500,6 @@ final class ConnectionRecoveryHandler_Tests: XCTestCase {
         handler.webSocketClient(mockChatClient.mockWebSocketClient, didUpdateConnectionState: .connecting)
 
         XCTAssertNotCall("syncLocalState(completion:)", on: mockChatClient.mockSyncRepository)
-        XCTAssertNil(mockChatClient.mockExtensionLifecycle.receivedIsReceivingEvents)
     }
 
     func test_webSocketStateUpdate_connecting_whenTimeout_whenNotRunning_shouldStartTimeout() {
@@ -510,7 +509,6 @@ final class ConnectionRecoveryHandler_Tests: XCTestCase {
         handler.webSocketClient(mockChatClient.mockWebSocketClient, didUpdateConnectionState: .connecting)
 
         XCTAssertNotCall("syncLocalState(completion:)", on: mockChatClient.mockSyncRepository)
-        XCTAssertNil(mockChatClient.mockExtensionLifecycle.receivedIsReceivingEvents)
     }
 
     func test_webSocketStateUpdate_connecting_whenTimeout_whenRunning_shouldNotStartTimeout() {
@@ -520,7 +518,6 @@ final class ConnectionRecoveryHandler_Tests: XCTestCase {
         handler.webSocketClient(mockChatClient.mockWebSocketClient, didUpdateConnectionState: .connecting)
 
         XCTAssertNotCall("syncLocalState(completion:)", on: mockChatClient.mockSyncRepository)
-        XCTAssertNil(mockChatClient.mockExtensionLifecycle.receivedIsReceivingEvents)
     }
 
     func test_webSocketStateUpdate_connected() {
@@ -531,7 +528,6 @@ final class ConnectionRecoveryHandler_Tests: XCTestCase {
 
         XCTAssertCall(RetryStrategy_Spy.Signature.resetConsecutiveFailures, on: mockRetryStrategy, times: 1)
         XCTAssertCall("syncLocalState(completion:)", on: mockChatClient.mockSyncRepository, times: 1)
-        XCTAssert(mockChatClient.mockExtensionLifecycle.receivedIsReceivingEvents == true)
     }
 
     func test_webSocketStateUpdate_connected_whenTimeout_shouldStopTimeout() {
@@ -542,7 +538,6 @@ final class ConnectionRecoveryHandler_Tests: XCTestCase {
 
         XCTAssertCall(RetryStrategy_Spy.Signature.resetConsecutiveFailures, on: mockRetryStrategy, times: 1)
         XCTAssertCall("syncLocalState(completion:)", on: mockChatClient.mockSyncRepository, times: 1)
-        XCTAssert(mockChatClient.mockExtensionLifecycle.receivedIsReceivingEvents == true)
     }
 
     func test_webSocketStateUpdate_disconnected_userInitiated() {
@@ -558,7 +553,6 @@ final class ConnectionRecoveryHandler_Tests: XCTestCase {
         XCTAssertNotCall(RetryStrategy_Spy.Signature.nextRetryDelay, on: mockRetryStrategy)
         XCTAssertNotCall("incrementConsecutiveFailures()", on: mockRetryStrategy)
         XCTAssertNotCall("syncLocalState(completion:)", on: mockChatClient.mockSyncRepository)
-        XCTAssert(mockChatClient.mockExtensionLifecycle.receivedIsReceivingEvents == false)
     }
 
     func test_webSocketStateUpdate_disconnected_systemInitiated() {
@@ -578,7 +572,6 @@ final class ConnectionRecoveryHandler_Tests: XCTestCase {
         XCTAssertCall(RetryStrategy_Spy.Signature.nextRetryDelay, on: mockRetryStrategy, times: 1)
         XCTAssertCall("incrementConsecutiveFailures()", on: mockRetryStrategy, times: 1)
         XCTAssertNotCall("syncLocalState(completion:)", on: mockChatClient.mockSyncRepository)
-        XCTAssert(mockChatClient.mockExtensionLifecycle.receivedIsReceivingEvents == false)
     }
 
     func test_webSocketStateUpdate_initialized() {
@@ -588,7 +581,6 @@ final class ConnectionRecoveryHandler_Tests: XCTestCase {
         handler.webSocketClient(mockChatClient.mockWebSocketClient, didUpdateConnectionState: .initialized)
 
         XCTAssertNotCall("syncLocalState(completion:)", on: mockChatClient.mockSyncRepository)
-        XCTAssertNil(mockChatClient.mockExtensionLifecycle.receivedIsReceivingEvents)
     }
 
     func test_webSocketStateUpdate_waitingForConnectionId() {
@@ -598,7 +590,6 @@ final class ConnectionRecoveryHandler_Tests: XCTestCase {
         handler.webSocketClient(mockChatClient.mockWebSocketClient, didUpdateConnectionState: .waitingForConnectionId)
 
         XCTAssertNotCall("syncLocalState(completion:)", on: mockChatClient.mockSyncRepository)
-        XCTAssertNil(mockChatClient.mockExtensionLifecycle.receivedIsReceivingEvents)
     }
 
     func test_webSocketStateUpdate_disconnecting() {
@@ -611,7 +602,6 @@ final class ConnectionRecoveryHandler_Tests: XCTestCase {
         )
 
         XCTAssertNotCall("syncLocalState(completion:)", on: mockChatClient.mockSyncRepository)
-        XCTAssertNil(mockChatClient.mockExtensionLifecycle.receivedIsReceivingEvents)
     }
 }
 
@@ -626,7 +616,6 @@ private extension ConnectionRecoveryHandler_Tests {
             webSocketClient: mockChatClient.mockWebSocketClient,
             eventNotificationCenter: mockChatClient.eventNotificationCenter,
             syncRepository: mockChatClient.mockSyncRepository,
-            extensionLifecycle: mockChatClient.extensionLifecycle,
             backgroundTaskScheduler: mockBackgroundTaskScheduler,
             internetConnection: mockInternetConnection,
             reconnectionStrategy: mockRetryStrategy,